+++ src/app/locales/en/utils/archive-browser-model.js
@@ -7,7 +7,9 @@ export default _.merge({}, BaseBrowserModel, {
   startingDownload: 'starting archive download',
   archiveList: 'Archives list',
   alreadyDeleting: 'Not available for archives that are being deleted.',
+  alreadyCancelling: 'Not available for archives that are being cancelled.',
   notAvailableForCreating: 'Not available for archives that are not built yet.',
+  notAvailableForIncomplete: 'Not available for archives that are not built successfully.',
   notAvailableForDetached: 'Not available in detached dataset.',
   fileActions: {
     archiveProperties: 'Properties',
@@ -21,8 +23,25 @@ export default _.merge({}, BaseBrowserModel, {
       multi: 'Delete archives',
       single: 'Delete archive',
     },
+    cancel: 'Cancel archivization',
     browseDip: 'Browse DIP',
   },
   archiveId: 'archive ID',
   selectedArchiveNoDip: 'Selected archive does not include DIP.',
+  cancelModal: {
+    header: 'Cancel archivization',
+    message: {
+      single: cancelModalMessage(false),
+      multi: cancelModalMessage(true),
+    },
+    yes: 'Yes',
+    no: 'No',
+    cancelling: 'cancelling archivization',
+  },
 });
+
+function cancelModalMessage(isMultiple) {
+  const selectedArchivesText = isMultiple ? ' for the {{archivesCount}} selected archives' : '';
+  const pluralSuffix = isMultiple ? 's' : '';
+  return `Are you sure you want to cancel the archivization process${selectedArchivesText}? The resulting archive${pluralSuffix} will be incomplete. Once cancelled, the process cannot be resumed.`;
+}